Reorganizes user components into seperate folder
